Adopt Phan
A static analyzer helps to detect some potential issues without even
running the code.  Phan appears to be particularly suitable for
seriously legacy code.  For now we configure it to report critical
issues only.

To avoid the dependency hell, we don't add it as composer development
requirement, but rather introduce Phive.

We also run the Phing target `phan` for CI.

Note that this raises the requirement to CMSimple_XH ≥ 1.7.0, since
earlier versions didn't use namespaced classes (and working around that
doesn't seem sensible).